Create Portfolio class
Usage:

    from portfolio import Portfolio
    p = Portfolio()

The performance of the portfolio will eventually live under p.performance
(which currently only has market days and whether they are open or not).

A specific account can also be invoked:

    p = Portfolio(account='Account Name')

This commit adds a Makefile, testing utilities, and tests for the above.

It also deletes the earlier ticker_prices function, which will be incorporated
within the portfolio class.